Directory renaming #44
