+++ src/app/features/models/capture/ui-model/current-event-processing-ui-model/activity-selecting-ui-model.ts
