+++ src/app/features/models/capture/ui-model/current-event-processing-ui-model/current-event-processing-ui-model.ts
